fix an issue with bloom by making sure a cvar is created for glsl to use (even if not directly used by the engine). this ensures there's no weirdness with defaults. some fixes/tweaks for the nolegacy builds. r_skybox works mid-map again. tweak the qc debugger to not do infinite loops so easily... rewrote chunks of the d3d11 renderer to support q3 shaders a little more correctly. now runs quake3 acceptably. add explicit culldistance option for terrain/.map maps. pvs will effectively cut off anything beyond this distance.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4982 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/server/progdefs.h
@@ -94,7 +94,7 @@ typedef struct nqglobalvars_s
 	vec3_t *input_movevalues;
 	float *input_buttons;
 	float *spawnparamglobals[NUM_SPAWN_PARMS];
-	float *serverid;
+	int *serverid;
 } globalptrs_t;
 
 #define P_VEC(v) (pr_global_struct->v)
